+++ java/ingest/ingest-runner/src/main/java/sleeper/ingest/job/NewIngestJobQueueConsumer.java
@@ -32,8 +32,8 @@ import sleeper.configuration.properties.instance.InstanceProperties;
 import sleeper.core.record.process.RecordsProcessedSummary;
 import sleeper.core.table.TableIndex;
 import sleeper.ingest.job.status.IngestJobStatusStore;
-import sleeper.ingest.task.NewIngestTask.MessageHandle;
-import sleeper.ingest.task.NewIngestTask.MessageReceiver;
+import sleeper.ingest.task.IngestTask.MessageHandle;
+import sleeper.ingest.task.IngestTask.MessageReceiver;
 import sleeper.io.parquet.utils.HadoopPathUtils;
 import sleeper.job.common.action.ActionException;
 import sleeper.job.common.action.MessageReference;
